Fixed game deletion bug

+++ app/controllers/games_controller.rb
@@ -21,9 +21,9 @@ class GamesController < ApplicationController
   end
 
   def destroy
-    game = Game.find(params[:id])
-    game.pieces.destroy_all
-    game.destroy
+    game = Game.find_by_id(params[:id])
+    game.pieces.destroy_all if game
+    game.destroy if game
     redirect_to root_path
   end
